Shadow fix
Best compromise i could find while playing with the settings. The sweet spot for shadow distance seems to be between 50 and 150 but the game scale is much larger than that. There is a way to use baked shadows on far objects but it does not work because our background is fully animated. Another option would be to use fog to hide ugly shadows in the distance but the result doesnt work well with our scene settings and the ortho camera.